Don't use %zu
Eventually we'll need to make sure that a) libroken's stdint.h defines
the max integer types, b) the libroken *printf()s can handle all the
standard length and conversion specifiers.

+++ lib/hx509/revoke.c
@@ -1151,7 +1151,7 @@ print_ocsp(hx509_context context, struct revoke_ocsp *ocsp, FILE *out)
 	    status = "element unknown";
 	}
 
-	fprintf(out, "\t%zu. status: %s\n", i, status);
+	fprintf(out, "\t%llu. status: %s\n", (unsigned long long)i, status);
 
 	fprintf(out, "\tthisUpdate: %s\n",
 		printable_time(ocsp->ocsp.tbsResponseData.responses.val[i].thisUpdate));
